rename package: execute->deployment
