+++ pkg/app/pipedv1/plugin/wait/deployment/server.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package execute
+package deployment
 
 import (
 	"context"
